% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_12C}
\alias{calc_12C}
\title{Identify the 12C peak for a compound and calculate its percentage vs all ions}
\usage{
calc_12C(raw_m)
}
\arguments{
\item{raw_m}{Raw data in melted form}
}
\description{
Identify the 12C peak for a compound and calculate its percentage vs all ions
}

