% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exp_coef}
\alias{exp_coef}
\title{Returns parameters of an exponential model}
\usage{
exp_coef(mod)
}
\arguments{
\item{mod}{An nls model. Right now should be an exponential model of the form y=Ae^(kx)}
}
\description{
Right now only works with two-factor exponential models
}

