% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_exp_guess}
\alias{generate_exp_guess}
\title{Generate guesses for A and k in exponential fit}
\usage{
generate_exp_guess(x, y)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}
}
\description{
Fits a linear least-squares fit to log-transformed exponential data
}
\details{
When this fails, it is a good sign that the data don't fit an exponential model. Often this would be either because there are too few data points (1 or 0) or some of the data points are negative. In that case, the function returns NULL. (Although I'm not convinced that is the optimal return value).
}

