% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_facet_formula}
\alias{generate_facet_formula}
\title{Generates expressions for facets}
\usage{
generate_facet_formula(exp.var.short, vert.var = "compound", omit = NULL,
  facet = "grid")
}
\description{
Right now used only by plot_timecourse.R
}

