% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_basic_timeseries_plot}
\alias{make_basic_timeseries_plot}
\title{Make basic timeseries plot}
\usage{
make_basic_timeseries_plot(df, save.plot = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{save.plot}{}
}
\description{
Make basic timeseries plot
}

