% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_timecourse}
\alias{plot_timecourse}
\title{Plots timecourses}
\usage{
plot_timecourse(df, exp.var, x.var = "time", y.var = "relative.ion.count",
  rep.var = "replicate", color.by = "sample.type", ignore.reps = TRUE,
  exp_pred = NULL, print.plot = TRUE, save.plot = FALSE, plot.fn, ...)
}
\arguments{
\item{df}{data frame containing %C12 data}

\item{exp.var}{character vector of experimental variables}

\item{x.var}{The x variable for the plot, usually time}

\item{rep.var}{The variable encoding replicates}

\item{color.by}{The variable by which to color the data. Usually treatment}

\item{exp_pred}{Data frame of exponential fit lines}

\item{print.plot}{CURRENTLY NOT IMPLEMENTED Whether to print the plot to screen. May not work outside of Rstudio}

\item{save.plot}{CURRENTLY NOT IMPLEMENTED Whether to save the plot.}

\item{plot.fn}{CURRENTLY NOT IMPLEMENTED Filename with which to save the plot}

\item{yvar}{The y variable for the plot, usually relative ion count}

\item{ingnore.reps}{Whether to ignore replicates as a vairable with which to separate the data}
}
\description{
Plots timecourses
}

