% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_timecourseOLD}
\alias{plot_timecourseOLD}
\title{Plot timecourses
Contains a nested if statement that is FROM THE DEVIL}
\usage{
plot_timecourseOLD(df, exp.var, x.var = "time",
  y.var = "relative.ion.count", treat.var = "treatment",
  rep.var = "replicate", distinguish.reps = FALSE, exp_pred,
  print.plot = TRUE, save.plot = FALSE, fn, ...)
}
\description{
Plot timecourses
Contains a nested if statement that is FROM THE DEVIL
}

