% Generated by roxygen2 (4.0.2): do not edit by hand
\name{safe_NLS}
\alias{safe_NLS}
\title{Calculates exponential models safely}
\usage{
safe_NLS(df, xvar = "time", yvar = "relative.ion.count")
}
\arguments{
\item{df}{A data frame}

\item{xvar}{The x variable for the exponential model}

\item{yvar}{The y variable for the exponential model}
}
\description{
Calculates exponential models safely
}

