% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exp_coef}
\alias{exp_coef}
\title{Returns parameters of an exponential model
Right now only works with two-factor exponential models}
\usage{
exp_coef(mod)
}
\description{
Returns parameters of an exponential model
Right now only works with two-factor exponential models
}

