% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_facet_formula}
\alias{generate_facet_formula}
\title{Generates expressions for facets}
\usage{
generate_facet_formula(exp.var.short, vert.var = "compound", omit = NULL,
  facet = "grid")
}
\arguments{
\item{exp.var.short}{Character vector of the names of the experimental variables. Currently there is some ambiguity about whether the time variable should be included in exp.var}

\item{vert.var}{Variable used for the vertical component of the facet #Describe this more clearly}

\item{omit}{Variables to remove}

\item{facet}{Type of facetting. "grid" or "wrap". Only grid is currently implemented.}
}
\description{
Right now used only by plot_timecourse.R
}

